+++ src/lib/contants.ts
@@ -1,0 +1,5 @@
+export const BRAND = {
+  title: "The Duck Council",
+  description:
+    "LLorem ipsum dolor sit amet, qui minim labore adipisicing minim sint cillum sint consectetur cupidatat.orem ipsum dolor sit amet, qui minim labore adipisicing minim sint cillum sint consectetur cupidatat.",
+};